+++ packages/portal/swagger-ui-web-component/src/utils.js
@@ -5,14 +5,20 @@ export const attributeValueToBoolean = (value) => {
 }
 
 // Yes, Swagger literally calls it "thing"
-export const escapeSwaggerThing = (str) => CSS.escape(str.trim().replace(/\s/g, '%20'))
+export const escapeSwaggerThing = (str) => {
+  return decodeURIComponent(str.trim().replace(/\s/g, '_'))
+}
 
-export const operationToSwaggerThingArray = (operation) => ([
-  'operations',
-  operation.tag ? escapeSwaggerThing(operation.tag) : 'default',
-  helpers.opId(operation, operation.path, operation.method),
-])
+export const operationToSwaggerThingArray = (operation) => {
+  return [
+    'operations',
+    operation.tag ? operation.tag : 'default',
+    operation.operationId ? operation.operationId : helpers.opId(operation, operation.path, operation.method),
+  ]
+}
 
-export const operationToSwaggerThingId = (operation) => escapeSwaggerThing(
-  operationToSwaggerThingArray(operation).join('-'),
-)
+export const operationToSwaggerThingId = (operation) => {
+  return escapeSwaggerThing(
+    operationToSwaggerThingArray(operation).join('-'),
+  )
+}